fix iplayer commands invoked using M-x
Sure, I only notice this when I have to restart my emacs.  But once is
annoying enough.  The problem was that running M-x iplayer would cause
an error from within execute-extended-command, within the process
sentinel.  After trying to debug this and generating a nifty
unbreakable infinite loop, I finally twigged that it was the
frame/window/buffer frobbing that was the problem.  Piling hack upon
hack, the answer is that we only want to do the frobbing once before
the first command in the command keys, and once after the execution
has finished.  Make it so.