don't startup by default in "show all" mode
my emacs takes a good couple of seconds merely to execute navigation
commands through the buffer with all the content, and isearch chews up
a *lot* of CPU.  It'd be nice to find out why and fix it, but in the
meantime start up in "BBC One" mode (by default; there's a customize
variable).

Also include emacs-style next/previous and refresh key bindings.